pyproject.toml
[tool.poetry]
name = "centribal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-baton = "^2.7.1"
psycopg2-binary = "^2.9.6"
django-cors-headers = "^3.14.0"
django-rest-authtoken = "^2.1.4"
django-rest-passwordreset = "^1.3.0"
phonenumbers = "^8.13.11"
django-phonenumber-field = "^7.1.0"
six = "^1.16.0"
django-simple-history = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
